Moved ParserGrammarFactory to grammar package
