+++ core/src/main/java/grammar/ParserGrammar.java
@@ -1,6 +1,5 @@
 package grammar;
 
-import automata.ParserGrammarFactory;
 import automata.PushdownAutomaton;
 
 import java.util.List;
