+++ core/src/test/java/automata/TestPushdownAutomata.java
@@ -1,6 +1,7 @@
 package automata;
 
 import grammar.LexerGrammar;
+import grammar.ParserGrammarFactory;
 import lexer.Lexer;
 import lexer.Token;
 import org.junit.Before;
